Move s3 stuff to s3browser
